Interface of SearchService (#254)
 - Create an interface to hold the legacy SearchService method
 - Added transfer class to isolate complicated termination processing
 - Add Name Setting for Sono
 - Moving your own class package and changing the test case operation
settings accordingly
