+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/service/search/IndexTypeTestCase.java
@@ -29,6 +29,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import junit.framework.TestCase;
 
 @RunWith(SpringJUnit4ClassRunner.class)
+@SuppressWarnings("deprecation")
 public class IndexTypeTestCase extends TestCase {
 
     @Test
